+++ lib/govuk_publishing_components.rb
@@ -11,8 +11,6 @@ require "govuk_publishing_components/presenters/content_breadcrumbs_based_on_tax
 require "govuk_publishing_components/presenters/services"
 require "govuk_publishing_components/presenters/select"
 require "govuk_publishing_components/presenters/meta_tags"
-require "govuk_publishing_components/presenters/taxonomy_navigation"
-require "govuk_publishing_components/presenters/curated_taxonomy_sidebar_links"
 require "govuk_publishing_components/presenters/content_item"
 require "govuk_publishing_components/presenters/translation_nav_helper"
 require "govuk_publishing_components/presenters/subscription_links_helper"