+++ spec/components/related_navigation_spec.rb
@@ -18,7 +18,7 @@ describe "Related navigation", type: :view do
   end
 
   it "renders nothing when no parameters given" do
-    assert_empty render_component({}).strip
+    assert_empty render_component(content_item: {}).strip
   end
 
   it "renders related content section when passed related items" do
@@ -26,7 +26,7 @@ describe "Related navigation", type: :view do
     content_item["links"] = construct_links(
       "ordered_related_items", "/apprenticeships", "Apprenticeships"
     )
-    render_component(content_item)
+    render_component(content_item: content_item)
 
     assert_select ".gem-c-related-navigation__main-heading", text: 'Related content'
     assert_select ".gem-c-related-navigation__section-link--other[href=\"/apprenticeships\"]", text: 'Apprenticeships'
@@ -37,7 +37,7 @@ describe "Related navigation", type: :view do
     content_item["links"] = construct_links(
       "related_guides", "/something-a-bit-like-this", "Some other guidance"
     )
-    render_component(content_item)
+    render_component(content_item: content_item)
 
     assert_select ".gem-c-related-navigation__sub-heading", text: 'Detailed guidance'
     assert_select ".gem-c-related-navigation__section-link[href=\"/something-a-bit-like-this\"]", text: 'Some other guidance'
@@ -48,7 +48,7 @@ describe "Related navigation", type: :view do
     content_item["links"] = construct_links(
       "topics", "/finding-a-job", "Finding a job", "topic"
     )
-    render_component(content_item)
+    render_component(content_item: content_item)
 
     assert_select ".gem-c-related-navigation__sub-heading", text: 'Explore the topic'
     assert_select ".gem-c-related-navigation__section-link[href=\"/finding-a-job\"]", text: 'Finding a job'
@@ -62,7 +62,7 @@ describe "Related navigation", type: :view do
       "Air quality statistics",
       "statistical_data_set"
     )
-    render_component(content_item)
+    render_component(content_item: content_item)
 
     assert_select ".gem-c-related-navigation__sub-heading", text: 'Statistical data set'
     assert_select ".gem-c-related-navigation__section-link[href=\"/air-quality-statistics\"]", text: 'Air quality statistics'
@@ -71,7 +71,7 @@ describe "Related navigation", type: :view do
   it "renders world locations section when passed world location items" do
     content_item = {}
     content_item["links"] = construct_links("world_locations", "/world/usa/news", "USA")
-    render_component(content_item)
+    render_component(content_item: content_item)
 
     assert_select ".gem-c-related-navigation__sub-heading", text: 'World locations'
     assert_select ".gem-c-related-navigation__section-link[href=\"/world/usa/news\"]", text: 'USA'
@@ -85,7 +85,7 @@ describe "Related navigation", type: :view do
       "The future of jobs and skills",
       "document_collection"
     )
-    render_component(content_item)
+    render_component(content_item: content_item)
 
     assert_select ".gem-c-related-navigation__sub-heading", text: 'Collection'
     assert_select ".gem-c-related-navigation__section-link[href=\"/government/collections/the-future-of-jobs-and-skills\"]", text: 'The future of jobs and skills'
@@ -99,7 +99,7 @@ describe "Related navigation", type: :view do
       "UK-China High-Level People to People Dialogue 2017",
       "topical_event"
     )
-    render_component(content_item)
+    render_component(content_item: content_item)
 
     assert_select ".gem-c-related-navigation__sub-heading", text: 'Topical event'
     assert_select ".gem-c-related-navigation__section-link[href=\"/government/topical-events/uk-china-high-level-people-to-people-dialogue-2017\"]", text: 'UK-China High-Level People to People Dialogue 2017'
@@ -115,7 +115,7 @@ describe "Related navigation", type: :view do
         }
       ]
     }
-    render_component(content_item)
+    render_component(content_item: content_item)
 
     assert_select ".gem-c-related-navigation__sub-heading--other", text: 'Elsewhere on the web'
     assert_select ".gem-c-related-navigation__section-link--other[href=\"https://www.thestudentroom.co.uk/content.php?r=5967-Repaying-your-student-loan\"][rel=\"external\"]", text: 'The Student Room repaying your student loan'
@@ -129,7 +129,7 @@ describe "Related navigation", type: :view do
       "A related contact",
       "contact"
     )
-    render_component(content_item)
+    render_component(content_item: content_item)
 
     assert_select ".gem-c-related-navigation__sub-heading--other", text: 'Other contacts'
     assert_select ".gem-c-related-navigation__section-link--other[href=\"/ye-olde-contact\"]", text: 'A related contact'
@@ -143,7 +143,7 @@ describe "Related navigation", type: :view do
       "Apprenticeships",
       "topic"
     )
-    render_component(content_item)
+    render_component(content_item: content_item)
 
     assert_select ".gem-c-related-navigation__nav-section[aria-labelledby]"
   end
@@ -153,7 +153,7 @@ describe "Related navigation", type: :view do
     %w(USA Wales Fiji Iceland Sweden Mauritius Brazil).each do |country|
       content_item["links"]["world_locations"] << { "title" => country }
     end
-    render_component(content_item)
+    render_component(content_item: content_item)
 
     assert_select ".gem-c-related-navigation__section-link[href=\"/world/wales/news\"]", text: 'Wales'
     assert_select ".gem-c-related-navigation__link.toggle-wrap", text: '+ 2 more'
@@ -166,7 +166,7 @@ describe "Related navigation", type: :view do
     %w(USA Wales Fiji Iceland Sweden Mauritius).each do |country|
       content_item["links"]["world_locations"] << { "title" => country }
     end
-    render_component(content_item)
+    render_component(content_item: content_item)
 
     assert_select ".gem-c-related-navigation__section-link[href=\"/world/wales/news\"]", text: 'Wales'
     assert_select ".gem-c-related-navigation__link.toggle-wrap", false, "Progressive disclosure should not display for only 1 link"
@@ -178,7 +178,7 @@ describe "Related navigation", type: :view do
       "ordered_related_items", "/apprenticeships", "Apprenticeships"
     )
     content_item["links"] = ordered_related_items
-    render_component(content_item)
+    render_component(content_item: content_item)
 
     assert_select ".gem-c-related-navigation__main-heading", text: 'Related content'
     assert_select ".gem-c-related-navigation__section-link--other[href=\"/apprenticeships\"]", text: 'Apprenticeships'
@@ -189,7 +189,7 @@ describe "Related navigation", type: :view do
     content_item["links"] = construct_links(
       "topics", "/apprenticeships", "Apprenticeships", "topic"
     )
-    render_component(content_item)
+    render_component(content_item: content_item)
 
     assert_select ".gem-c-related-navigation__nav-section ul[data-module='track-click']"
     assert_select ".gem-c-related-navigation__section-link[data-track-category='relatedLinkClicked']"