+++ spec/lib/presenters/related_navigation_helper_spec.rb
@@ -1,15 +1,15 @@
 require "spec_helper"
 
 RSpec.describe GovukPublishingComponents::Presenters::RelatedNavigationHelper do
-  def payload_for(schema, content_item)
+  def payload_for(schema, content_item, context = nil)
     example = GovukSchemas::RandomExample.for_schema(frontend_schema: schema) do |payload|
       payload.merge(content_item)
     end
 
-    described_class.new(example).related_navigation
+    described_class.new(content_item: example, context: context).related_navigation
   end
 
-  describe "#related-navigation-sidebar" do
+  describe "#related_navigation" do
     it "can handle randomly generated content" do
       expect { payload_for("placeholder", {}) }.to_not raise_error
     end
@@ -199,7 +199,7 @@ RSpec.describe GovukPublishingComponents::Presenters::RelatedNavigationHelper do
 
     it "handles ordered related items that aren't tagged to a mainstream browse page" do
       example = GovukSchemas::Example.find("guide", example_name: "single-page-guide")
-      payload = described_class.new(example).related_navigation
+      payload = described_class.new(content_item: example).related_navigation
       expected = [
         { text: "Travel abroad", path: "/browse/abroad/travel-abroad" },
         { text: "Arriving in the UK", path: "/browse/visas-immigration/arriving-in-the-uk" },
@@ -245,6 +245,93 @@ RSpec.describe GovukPublishingComponents::Presenters::RelatedNavigationHelper do
       expect(payload["related_contacts"]).to eql(
         [{ path: "/foo", text: "Foo" }]
       )
+    end
+
+    it "returns live taxons" do
+      payload = payload_for("placeholder",
+                            "details" => {
+                              "external_related_links" => []
+                            },
+                            "links" => {
+                              "taxons" => [
+                                {
+                                  "title" => "Taxon B",
+                                  "base_path" => "/taxon-b",
+                                  "content_id" => "82a84770-ec89-4f75-8335-4ff78d84d97d",
+                                  "document_type" => "taxon",
+                                  "description" => "The B taxon.",
+                                  "phase" => "live",
+                                  "locale" => "en",
+                                },
+                                {
+                                  "title" => "Taxon A",
+                                  "base_path" => "/taxon-a",
+                                  "content_id" => "c8743ebd-ceb3-493b-b66a-e3bb4d30b7be",
+                                  "document_type" => "taxon",
+                                  "description" => "The A taxon.",
+                                  "phase" => "live",
+                                  "locale" => "en",
+                                },
+                                {
+                                  "title" => "Taxon C",
+                                  "base_path" => "/taxon-c",
+                                  "content_id" => "6271264a-72fa-4c12-aaa2-ab3071d3d133",
+                                  "document_type" => "taxon",
+                                  "description" => "The C taxon.",
+                                  "phase" => "draft",
+                                  "locale" => "en",
+                                },
+                              ],
+                            })
+
+      expect(payload["topics"]).to eql(
+        [
+          { path: "/taxon-b", text: "Taxon B" },
+          { path: "/taxon-a", text: "Taxon A" },
+        ]
+       )
+    end
+
+    context 'for a sidebar' do
+      subject(:payload) { payload_for('placeholder', {}, :sidebar) }
+
+      it 'only includes collections, guides and related items' do
+        expect(payload).to include(
+          'collections',
+          'related_guides',
+          'related_items',
+        )
+
+        expect(payload).to_not include(
+          'related_contacts',
+          'related_external_links',
+          'statistical_data_sets',
+          'topical_events',
+          'topics',
+          'world_locations',
+        )
+      end
+    end
+
+    context 'for a footer' do
+      subject(:payload) { payload_for('placeholder', {}, :footer) }
+
+      it 'only includes contacts external links, statistical datasets, topical events, topics and world locations' do
+        expect(payload).to include(
+          'related_contacts',
+          'related_external_links',
+          'statistical_data_sets',
+          'topical_events',
+          'topics',
+          'world_locations',
+        )
+
+        expect(payload).to_not include(
+          'collections',
+          'related_guides',
+          'related_items',
+        )
+      end
     end
   end
 end