+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 
 ## Unreleased
 
+* Retires the taxonomy navigation component (PR #606)
+* Introduces a new contextual footer component (PR #606)
 * Use ERB comments not HTML comments in template markup (PR #667)
 
 ## 12.20.0